Updates to match Parser functionality
- Added BSON and MSGPack support
- Added AutoDetect support (via Accept header)
- Added ResponderException and made most types throw it (QueryStr doesn't throw any in Parser, so it also doesn't in Responder)
- Updated composer.json to use the same package versions as Parser
- Updated Travis test matrix to match Parser
- Various minor/cosmetic changes to bring Responder up to date with Parser

The idea here was to be able to use Parser and Responder together to build APIs that accept and return data in the same variety of formats.  This is useful because API clients can then communicate using whichever format(s) they are most comfortable using, in both directions, bypassing the need to include their own extra parser(s)/formatter(s) to handle extraneous formats.

Install Extensions For Testing

Test various combinations of extensions installed or not:
- No extensions
- Just MSGPack
- Just Mongo
- Just MongoDB
- MSGPack and MongoDB

Because APT is unreliable for this, download and compile extensions directly in Travis jobs themselves.